# CUDA 11.1: Wrap the new `nvrtcGetCUBIN` API