--- a/cupy_backends/cuda/libs/nvrtc.pxd
+++ b/cupy_backends/cuda/libs/nvrtc.pxd
@@ -15,6 +15,7 @@ cpdef check_status(int status)
 
 cpdef tuple getVersion()
 
+
 ###############################################################################
 # Program
 ###############################################################################
@@ -24,6 +25,7 @@ cpdef intptr_t createProgram(unicode src, unicode name, headers,
 cpdef destroyProgram(intptr_t prog)
 cpdef compileProgram(intptr_t prog, options)
 cpdef bytes getPTX(intptr_t prog)
+cpdef bytes getCUBIN(intptr_t prog)
 cpdef unicode getProgramLog(intptr_t prog)
 cpdef addAddNameExpression(intptr_t prog, str name)
 cpdef str getLoweredName(intptr_t prog, str name)