--- a/cupy_backends/hip/cupy_hiprtc.h
+++ b/cupy_backends/hip/cupy_hiprtc.h
@@ -42,6 +42,14 @@ nvrtcResult nvrtcGetPTX(nvrtcProgram prog, char* code) {
     return hiprtcGetCode(prog, code);
 }
 
+nvrtcResult nvrtcGetCUBINSize(...) {
+    return HIPRTC_ERROR_COMPILATION;
+}
+
+nvrtcResult nvrtcGetCUBIN(...) {
+    return HIPRTC_ERROR_COMPILATION;
+}
+
 nvrtcResult nvrtcGetProgramLogSize(nvrtcProgram prog, std::size_t* logSizeRet) {
     return hiprtcGetProgramLogSize(prog, logSizeRet);
 }