--- a/cupy_backends/stub/cupy_nvrtc.h
+++ b/cupy_backends/stub/cupy_nvrtc.h
@@ -40,6 +40,14 @@ nvrtcResult nvrtcGetPTX(...) {
     return NVRTC_SUCCESS;
 }
 
+nvrtcResult nvrtcGetCUBINSize(...) {
+    return NVRTC_SUCCESS;
+}
+
+nvrtcResult nvrtcGetCUBIN(...) {
+    return NVRTC_SUCCESS;
+}
+
 nvrtcResult nvrtcGetProgramLogSize(...) {
     return NVRTC_SUCCESS;
 }