**UPDATE**: this PR only adds Cython wrappers for `nvrtcGetCUBIN`, and the migration strategy to this new API is to be discussed.

This PR changes the NVRTC default for CUDA 11.1+ from compiling PTX to compiling CUBIN, bypassing the driver JIT step when the module is loaded. This could potentially improve the overall performance (to be verified) in terms of both reducing overall time and enabling better code generation.

For CUDA 9.2 - 11.0, this PR does not do anything.

TODO:
- [x] Run some benchmarks to see if we get some speedup.