@leofang 2021-01-18
2719bff is a demo to show the minimal change needed to make this work on CUDA 11.1. But we need to retain backward compatibility for older CUDA as well as HIP, so some more work will be required.