@leofang 2021-01-19
Looks like this new feature actually increases the compilation overhead...😞 

Below I run two test files for different purposes:
- `test_raw.py`: Tests in this file do not involve the cubin cache on disk, so it gives a measure for total compilation time
- `test_fft.py`: Kernels used in this test are cached on disk, so the first run involves compiling, and the second run involves loading from disk (which in theory would involve driver JIT)

Timing | `test_raw.py` | `test_fft.py` (1st run) | `test_fft.py` (2nd run)
------|--------------|-------------|--------
before this PR | 44.8 s | 36.4 s | 10.7 s
with this PR | 54.8 s | 45.3 s | 10.8 s

So we can see 1. the compile time increases, 2. the load time does not decrease. Not sure what I'm missing to properly utilize this new feature...