@leofang 2021-01-22
OK I see where I got it wrong. The JIT'd binaries are cached by the driver, and I forgot to disable it. By setting `CUDA_CACHE_DISABLE=1` I got a timing making better sense:

Timing | `test_raw.py` | `test_fft.py` (1st run) | `test_fft.py` (2nd run)
------|--------------|-------------|--------
before this PR | 52.4 s | 46.3 s | 20.9 s
with this PR | 55.1 s | 45.2 s | 10.8 s 

1. The increase of compile time is mild, and likely only because Jitify still generates PTX (in `test_raw.py`)
2. The driver JIT time is not negligible and was saved because CUDA did the cache for us

@takagi How about we switch to `getCUBIN()` from `getPTX()` when we detect a user sets `CUDA_CACHE_DISABLE=1`?