@takagi 2021-01-25
I think switching when `CUDA_CACHE_DISABLE=1` is okay. BTW, CuPy's kernel cache on the disk will be changed from PTX to cubin, then what happens if the GPU on the machine is replaced with another and the cached cubin is loaded??