@leofang 2021-01-27
Hi @takagi I am still wondering if `getCUBIN()` should always be enable on CUDA 11.1+. By doing so we shift the cache responsibility from CUDA driver to CuPy, but we already do this anyway.

For invalidating the CUBIN when using a different GPU is in use, I think it's already done when generating CuPy's cache key. I will add CUDA version to the key to further disambiguate it. (By the way, on HIP we also cache binary files (hasco).)

Let me know what you think and I'll proceed accordingly 🙂