@takagi 2021-01-28
Thanks for explaining! I also checked CuPy's cache key and looks having enough information to invalidate the CUBIN properly.

BTW, I read through the changes again and now I think using `nvrtcGetCUBIN` only for the case `CUDA_CACHE_DISABLE=1` with adding extra branches doesn't make us much happier. How about just adding the wrappers for `nvrtcGetCUBIN` and cleaning some codes at this time?