@leofang 2021-01-28
> BTW, I read through the changes again and now I think using `nvrtcGetCUBIN` only for the case `CUDA_CACHE_DISABLE=1` with adding extra branches doesn't make us much happier. 

Yes, this is what I was wondering too. 

> How about just adding the wrappers for `nvrtcGetCUBIN` and cleaning some codes at this time?

Did you mean we don't use `getCUBIN()` in `cupy/cuda/compiler.py` at all, or you meant something else? I am happy to do either way, just need to clarify.