@takagi 2021-01-29
> Did you mean we don't use getCUBIN() in cupy/cuda/compiler.py at all, or you meant something else?

Yes, what I meant is the former. Of course, I always respect your highly productive works!