@leofang 2021-01-29
Thanks, @takagi. I cleaned up `compiler.py` and `test_raw.py`, and left the Cython wrappers and `_get_arch()` untouched (https://github.com/cupy/cupy/pull/4558#discussion_r559974610). PTAL.