@leofang 2021-02-01
Thanks, @takagi! I'll open an issue to discuss when to turn on `getCUBIN()`.